refactor: rename test_deps.ts -> dev_deps.ts
